% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_from_immunity.R
\name{aggregate_vacc_activities}
\alias{aggregate_vacc_activities}
\title{Aggregate matching vaccination activities across age groups}
\usage{
aggregate_vacc_activities(vacc_act, n_digits = 10)
}
\arguments{
\item{vacc_act}{vip_vacc_activities object to be aggregated}

\item{n_digits}{number of digits to which coverage is rounded
before coverages from different activities are matched.}
}
\value{
vip_vacc_activities object containing the same vaccination
activities as the input object, but where possible aggregated
into fewer lines
}
\description{
The vaccination activities contained in the input dataset are
matched by region, year, coverage and targeting method, and for
any that match on these variables, the age range will be
compressed into a (or several) consecutive age range, such that
the same information is coded in fewer lines.
}
\author{
Tini Garske
}
