% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{apply_vacc}
\alias{apply_vacc}
\title{Function to apply a single vaccination activity to the population}
\usage{
apply_vacc(pop_df, vacc_act)
}
\arguments{
\item{pop_df}{a population dataframe object such as created by
function setup_population}

\item{vacc_act}{details of the vaccination activity to be
implemented as a named list containing the elements "year",
"age_first", "age_last", "country", "coverage".}
}
\value{
The original pop_df dataframe with column "immunity"
updated to reflect the vaccination activity applied
}
\description{
Function to apply a single vaccination activity to the population
}
