% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popim_vacc_activities.R
\name{as_vacc_activities}
\alias{as_vacc_activities}
\title{Generate a \code{popim_vacc_activites} object from a dataframe}
\usage{
as_vacc_activities(df)
}
\arguments{
\item{df}{dataframe to be converted to class
\code{popim_vacc_activities}}
}
\value{
an object of class \code{popim_vacc_activities}
}
\description{
Checks if the dataframe is suitable (i.e., contains appropriate
columns and data ranges), and if so converts it to a
\code{popim_vacc_activities} object and returns this.
}
\details{
The input dataframe has to have at least the columns \code{region},
\code{year}, \code{age_first}, \code{age_last}, \code{coverage}, \code{doses} and
\code{targeting}. Any further columns are simply carried over into the
\code{popim_vacc_activities} object.
}
\seealso{
\code{\link[=popim_vacc_activities]{popim_vacc_activities()}} for details of the S3 class.
}
\author{
Tini Garske
}
