% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_coverage.R
\name{calc_new_coverage}
\alias{calc_new_coverage}
\title{Function to combine two coverages for the same cohort}
\usage{
calc_new_coverage(cov1, cov2, skew = 0)
}
\arguments{
\item{cov1}{first coverage value as a proportion - needs to be
between 0 and 1}

\item{cov2}{second coverage value as a proportion - needs to be
between 0 and 1}

\item{skew}{integer to determine the assumption used to combine
coverages. Valid values are 1, 0 and -1. For skew = 0 (the
default), allocation of vaccine is random within the
population, therefore the resulting coverage will be smaller
than the sum of the coverages. For skew = 1, there is a 100\%
correlation between who will get the vaccine in either
vaccination activitiy, the resulting coverage is simply the
larger of the two inputs. For skew = -1, doses are targeted at
unvaccinated people, resulting in the sum of both coverages,
though capped by 1 (full coverage).}
}
\value{
combined coverage
}
\description{
This function to calculate the resulting coverage if coverages of
two separate vaccination activities are combined, and returns this
single value.
}
