% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_immunity.R
\name{calc_new_immunity}
\alias{calc_new_immunity}
\title{Function to combine two coverages for the same cohort}
\usage{
calc_new_immunity(coverage, prev_immunity, targeting)
}
\arguments{
\item{coverage}{coverage value of the new campaign as a proportion
- needs to be between 0 and 1.}

\item{prev_immunity}{pre-existing immunity of the cohort to be
vaccinated as a proportion. Can be a vector (e.g. for
calculation of resulting immunity in the same cohort forward in
time). All entries need to be between 0 and 1.}

\item{targeting}{char to determine the assumption used to combine
coverages. Valid values are "random", "correlated", "targeted".
For targeting = "random" (the default), allocation of vaccine is
random within the population, therefore the resulting coverage
will be smaller than the sum of the coverages. For targeting =
"correlated", there is a 100\% correlation between who will get
the vaccine in either vaccination activitiy, the resulting
coverage is simply the larger of the two inputs. For targeting =
"targeted", doses are targeted at unvaccinated people,
resulting in the sum of both coverages, though capped by 1
(full coverage).}
}
\value{
A scalar: combined coverage
}
\description{
This function calculates the resulting coverage if coverages of
two separate vaccination activities are combined, and returns this
single value.
}
