% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_immunity.R
\name{calc_pop_immunity}
\alias{calc_pop_immunity}
\title{Aggregate the population immunity over age}
\usage{
calc_pop_immunity(pop)
}
\arguments{
\item{pop}{A \code{popim_population} object for which the population size and
immunity will be aggregated over age.}
}
\value{
A dataframe containing the \code{popim_population} aggregated by age.
}
\description{
Calculate the overall population immunity (aggregating over age)
from the supplied \code{popim_population} object.
}
\examples{
## set up population and vaccination activities:
pop <- popim_population(region = "UK", year_min = 2000, year_max = 2005,
                        age_min = 0, age_max = 10)
vacc <- popim_vacc_activities(region = "UK", year = c(2001, 2002),
                              age_first = 0, age_last = 0,
                              coverage = 0.8, doses = NA,
                              targeting = "random")

## update the population immunity based on the vaccination activities:
pop <- apply_vacc(pop, vacc)

## calculate the population immunity aggregated across ages:
pop_aggregated <- calc_pop_immunity(pop)

}
\author{
Tini Garske
}
