% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_population.R
\name{convert_df_to_pop}
\alias{convert_df_to_pop}
\title{Convert a dataframe to a vip_population object}
\usage{
convert_df_to_pop(df)
}
\arguments{
\item{df}{a dataframe with at least columns region, age, year and pop_size.}
}
\value{
an object of class vip_populaiton
}
\description{
\code{convert_df_to_pop} converts a suitable dataframe to a vip
population object and returns this object.
}
\details{
The input dataframe has to have at least the columns region, age,
and year. The output vip_population object is generated via
expand.grid to have consecutive year and age ranges that are
identical for all regions.

If the input dataframe contains a column \code{pop_size}, this must be
numeric and non-negative. If it is missing, this column is
generated and initialised to NA.

If the input dataframe contains a colum \code{immunity}, this must be
numeric, with values between 0 and 1. If it is missing, this
column is generated and initialised to 0.

Any further colunms are simply carried over into the vip_population object.
}
\author{
Tini Garske
}
