% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_formats.R
\name{convert_pop_to_wide}
\alias{convert_pop_to_wide}
\title{Converts from the standard vip_population format to the "wide" format}
\usage{
convert_pop_to_wide(pop_long)
}
\arguments{
\item{pop_long}{object of class "vip_population"}
}
\value{
object of class "vip_pop_wide"
}
\description{
Takes a vip_population object and converts it to the "wide" format
generated by vip_pop_wide() that is based on 3d arrays for region
/ year / age.
}
\author{
Tini Garske
}
