% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{coverage_from_doses}
\alias{coverage_from_doses}
\title{Calculate the coverage of a vaccination activity achieved with a
given number of doses}
\usage{
coverage_from_doses(pop_df, doses, region, year, age_first = 0, age_last = Inf)
}
\arguments{
\item{pop_df}{object of class popim_population such as created by
function \code{popim_population}}

\item{doses}{number of vaccine doses available for the vaccination
activity}

\item{region}{region to be targeted}

\item{year}{year of vaccination activity}

\item{age_first}{youngest age group to be targeted}

\item{age_last}{oldest age group to be targeted}
}
\value{
coverage: a number between 0 and 1 indicating what
proportion of the target population can be covered with the
given number of doses.
}
\description{
The coverage for a vaccination activity is calculated as the
proportion of the target population that can be vaccinated by the
given number of doses. The target population is the sum of the
population of the targeted age groups in the targeted region(s).
}
\author{
Tini Garske
}
