% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{doses_from_coverage}
\alias{doses_from_coverage}
\title{Calculate the number of doses needed to achieve a given coverage
in a vaccination activity}
\usage{
doses_from_coverage(
  pop_df,
  coverage,
  region,
  year,
  age_first = 0,
  age_last = Inf
)
}
\arguments{
\item{pop_df}{object of class popim_population such as created by
function \code{popim_population}}

\item{coverage}{proportion of the target population to be covered,
0 <= coverage <= 1}

\item{region}{region to be targeted}

\item{year}{year of vaccination activity}

\item{age_first}{youngest age group to be targeted}

\item{age_last}{oldest age group to be targeted}
}
\value{
doses: a non-negative number between indicating how many
doses will be needed to achieve the given coverage in the
target population.
}
\description{
The number of doses needed to achieve the given coverage is
calculated as coverage * target population. The target population
is the sum of the population of the targeted age groups in the
targeted region(s).
}
\author{
Tini Garske
}
