% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_from_immunity.R
\name{get_all_ages}
\alias{get_all_ages}
\title{Expand the age range given by age_first and age_last into an
integer vector giving each individual age group}
\usage{
get_all_ages(va)
}
\arguments{
\item{va}{object of class "popim_vacc_activities"}
}
\value{
integer vector of variable length
}
\description{
If the "popim_vacc_activities" object contains more than one row the
rows are assumed to belong to the same campaign, but detail
different age groups. The function will expand the age ranges to a
vector that gives every single age cohort targeted.
}
\author{
Tini Garske
}
