% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_from_immunity.R
\name{get_consecutive_range}
\alias{get_consecutive_range}
\title{Summarise the individual ages into one or more consecutive age ranges}
\usage{
get_consecutive_range(ages)
}
\arguments{
\item{ages}{non-negative integer vector of individual ages to be
targeted in a vaccination activity.}
}
\value{
list of age ranges: The list elements are integer vectors
of length 2, giving the age_first and age_last of the age
range into which the ages can be aggregated. The length of the
list depends on the complexity of the input ages.
}
\description{
Summarise the individual ages into one or more consecutive age ranges
}
\author{
Tini Garske
}
