% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.scalar}
\alias{is.scalar}
\title{Function to test if the argument is scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{argument to be tested for scalar-ness.}
}
\value{
logical - TRUE for a scalar, FALSE otherwise.
}
\description{
This function tests whether the argument is a scalar, i.e. not a
list or dataframe, is a vector of length 1 and is numeric. Returns
TRUE if this is the case, FALSE otherwise.
}
