% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_population.R
\name{is_population}
\alias{is_population}
\title{Function to test if the argument is a valid population dataframe}
\usage{
is_population(x)
}
\arguments{
\item{x}{argument to be tested if it fulfils the requirements for
a population dataframe.}
}
\value{
logical
}
\description{
This function tests whether the supplied argument is a valid
population dataframe such as might be generated by the function
"setup_population".  This currently only tests the structure, not
the content of the dataframe.
}
\author{
Tini Garske
}
