% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_scalar_integer}
\alias{is_scalar_integer}
\title{Function to test if the argument is a scalar integer}
\usage{
is_scalar_integer(x)
}
\arguments{
\item{x}{argument to be tested for scalar integer-ness.}
}
\value{
logical - TRUE for a scalar integer, FALSE otherwise.
}
\description{
This function tests whether the argument is a scalar integer,
i.e., a scalar as defined by is.scalar, and an integer.
}
\author{
Tini Garske
}
