% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Function to test if the argument is a whole number}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{argument to be tested for whole-ness.}

\item{tol}{tolerance for deviation from a whole number, by default
set to the square root of Machine precision.}
}
\value{
logical (of the same dimensions as the input)
}
\description{
Doing what one would expect is.integer to do...
}
\author{
R-help for is.integer
}
