% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_population.R
\name{new_population}
\alias{new_population}
\title{Low-level constructor of an object of the "vip_population" class}
\usage{
new_population(
  year_min = integer(),
  year_max = integer(),
  age_min = integer(),
  age_max = integer()
)
}
\arguments{
\item{year_min}{integer, first year to be considered.}

\item{year_max}{integer, last year to be considered.}

\item{age_min}{integer, youngest age to be considered.}

\item{age_max}{integer, oldest age to be considered.}
}
\value{
S3 object of class "vip_population": a dataframe with
columns year, age, cohort, immunity where year and age cover
the ranges given by the input parameters. Cohort = year - age
and gives the year of birth. It is redundant but included for
ease of handling. Immunity is initialised as 0 throughout the
whole population.
}
\description{
This function is intended for developer use only and therefore
doesn't perform any checks on input parameters.
}
\author{
Tini Garske
}
