% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_immunity}
\alias{plot_immunity}
\title{Plot the vaccine-derived immunity of a \code{popim_population} object}
\usage{
plot_immunity(pop_df)
}
\arguments{
\item{pop_df}{\code{popim_population} object such as created by
\code{\link[=popim_population]{popim_population()}}}
}
\value{
ggplot object
}
\description{
Plot the vaccine-derived immunity of a \code{popim_population} object
}
\details{
The population is displayed in a grid showing the cohorts
through time. Time is shown on the x-axis, age on the y-axis,
such that a particular cohort tracks along diagonally from
bottom left to top right. If there are several regions, these
are shown as separate facets. The colour in each cell
corresponds to the proportion of the cohort that is immune.
}
\author{
Tini Garske
}
