% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population.R
\name{plot_population}
\alias{plot_population}
\title{Function to plot the vaccine-derived immunity of a population}
\usage{
plot_population(pop_df)
}
\arguments{
\item{pop_df}{population dataframe object such as created by
function setup_population}
}
\value{
ggplot object
}
\description{
Function to plot the vaccine-derived immunity of a population
}
\author{
Tini Garske
}
