% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{plot_immunity}
\alias{plot_pop_size}
\title{Plot the immunity or population size of a \code{popim_population} object}
\usage{
plot_immunity(pop)

plot_pop_size(pop)
}
\arguments{
\item{pop}{\code{popim_population} object such as created by
\code{\link[=popim_population]{popim_population()}}.}
}
\value{
A ggplot object.
}
\description{
Plot the immunity or population size of a \code{popim_population} object
}
\details{
The population is displayed in a grid showing the cohorts
through time. Time is shown on the x-axis, age on the y-axis,
such that a particular cohort tracks along diagonally from
bottom left to top right. If there are several regions, these
are shown as separate facets.

The colour in each cell corresponds to:
\itemize{
\item for \code{plot_immunity()}: the proportion of each cohort that is
immune, therefore varying between 0 and 1.
\item for \code{plot_pop_size()}: the size of each cohort.
}

As the returned object is a regular \code{ggplot} object, it can be
further modified with the ususal \code{ggplot2} syntax.
}
\examples{
## set up population and vaccination activities:
pop <- popim_population(region = "UK", year_min = 2000, year_max = 2005,
                        age_min = 0, age_max = 10)
vacc <- popim_vacc_activities(region = "UK", year = c(2001, 2002),
                              age_first = 0, age_last = 0,
                              coverage = 0.8, doses = NA,
                              targeting = "random")

## update the population immunity based on the vaccination activities:
pop <- apply_vacc(pop, vacc)

## plot the population size by age and time:
plot_pop_size(pop)

## plot the population immunity by age and time:
plot_immunity(pop)

}
\author{
Tini Garske
}
