% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popim-package.R
\docType{package}
\name{popim-package}
\alias{popim}
\alias{popim-package}
\title{popim: POPulation IMmunity: Run a Demographic Model of Vaccine Exposure Over Time}
\description{
Tools for setting up an age-structured population, applying vaccination activities to it, and tracking vaccine-induced immunity through time.
}
\author{
\strong{Maintainer}: Tini Garske \email{t.garske@imperial.ac.uk} (\href{https://orcid.org/0000-0002-8952-4710}{ORCID})

}
\keyword{internal}
