% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_population.R
\name{read_population}
\alias{read_population}
\title{Reading population data from a .csv file}
\usage{
read_population(file)
}
\arguments{
\item{file}{Name of the file from which the population data are to
be read. If it does not contain an absolute path, the file
name is relative to the current working directory.}
}
\value{
object of class "vip_population", a dataframe with one row
per birth cohort/year/region, with columns "region", "year",
"age", "cohort", "immunity", "pop_size".
}
\description{
Reads a population data from a .csv file, checks if
the data fulfils the requirements for a "vip_population"
object, and if so returns this object.
}
\details{
The requirements are that the data contain the columns "region",
"age", "year". Columns "pop_size" and "immunity" are optional. The
columns will be coerced to character, integer, integer, numeric,
double, respectively.  If the column "immunity" exists, it will be
used to initialise the immunity data, if it doesn't exist,
immunity will be initialised to zero. If colunn pop_size doesn't
exist, it will be initialised to NA.

0 <= age
0 <= immunity <= 1
0 <= pop_size
}
\author{
Tini Garske
}
