% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_vacc_activities.R
\name{read_vacc_activities}
\alias{read_vacc_activities}
\title{Reading vaccination activities from a .csv file}
\usage{
read_vacc_activities(file)
}
\arguments{
\item{file}{Name of the file from which the vaccination activities
are to be read from. If it does not contain an absolute path,
the file name is relative to the current working directory.}
}
\value{
object of class "vip_vacc_activities", a dataframe with
one row per vaccination activity, with columns "year",
"age_first", "age_last", "coverage", "target".
}
\description{
Reads a list of vaccination activities from a .csv file, checks if
the data fulfils the requirements for a "vip_vacc_activities"
object, and if so returns this object.
}
\details{
The requirements are that the data contain the columns "year",
"age_first", "age_last", "coverage" and "target", of types
integer, integer, integer, double, character, respectively.

0 <= age_first <= age_last
0 <= coverage <= 1
target must be one of "random", "correlated", "targeted".
}
\author{
Tini Garske
}
