% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unwpp_countries}
\alias{unwpp_countries}
\title{List of countries in the UN WPP data 2024}
\format{
\code{unwpp_countries}
A data frame with 237 rows and 4 columns.
\describe{
\item{Location}{Name of the country}
\item{ISO3_code}{3-letter ISO country code}
\item{ISO2_code}{2-letter ISO country code}
\item{SDMX_code}{SDMX country code}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/CSV/}
}
\usage{
unwpp_countries
}
\description{
A list of the countries and country codes available in the UNWPP
population size data by single age in .csv format, useful for
initialising \code{popim_population} objects. The .csv files can be
downloaded at
\url{https://population.un.org/wpp/Download/Standard/CSV/}.
}
\keyword{datasets}
