% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{apply_vacc}
\alias{apply_vacc}
\title{Apply vaccination activities to a \code{popim_population} object}
\usage{
apply_vacc(pop, vacc)
}
\arguments{
\item{pop}{An object of class \code{popim_population} such as
created by \code{\link[=popim_population]{popim_population()}}.}

\item{vacc}{An object of class \code{popim_vacc_activities} such as
created by \code{\link[=popim_vacc_activities]{popim_vacc_activities()}}.}
}
\value{
The input \code{popim_population} class object \code{pop} with
updated column \code{immunity} to reflect the vaccination
activities supplied in \code{vacc}.
}
\description{
Takes the vaccination activities listed in the input object
vaccs_df (a data.frame object of class popim_vacc_activities), and
applies them to the popim_population object pop, then returns
that updated object.
}
\author{
Tini Garske
}
