% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{apply_vacc}
\alias{apply_vacc}
\title{Function to apply a single vaccination activity to the population}
\usage{
apply_vacc(pop_df, year, age_first, age_last, coverage, skew = 0)
}
\arguments{
\item{pop_df}{population dataframe object such as created by
function setup_population}

\item{year}{year of the vaccination activity}

\item{age_first}{age of the youngest age group targeted}

\item{age_last}{age of the oldest age group targeted}

\item{coverage}{proportion of the population to be vaccinated in
the activity}

\item{skew}{parameter to determine how successive activities are
targeted: see function calc_new_coverage for more details}
}
\value{
pop_df: the supplied population dataframe object with
updated immunity to reflect the vaccination activity.
}
\description{
Function to apply a single vaccination activity to the population
}
\author{
Tini Garske
}
