% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vacc.R
\name{apply_vaccs}
\alias{apply_vaccs}
\title{Function to apply vaccination activities to a population}
\usage{
apply_vaccs(pop_df, vaccs_df)
}
\arguments{
\item{pop_df}{object of class vip_population such as created by
function 'vip_population'}

\item{vaccs_df}{object of class 'vip_vacc_activities' such as
created by reading from file with function
'read_vacc_activities'}
}
\value{
pop_df: the supplied population dataframe object with
updated immunity to reflect the vaccination activities
supplied in vaccs_df.
}
\description{
Takes the vaccination activities listed in the input object
vaccs_df (a data.frame object of class vip_vacc_activities), and
applies them to the vip_population object pop_df, then returns
that updated object.
}
\author{
Tini Garske
}
