% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_immunity.R
\name{calc_pop_immunity}
\alias{calc_pop_immunity}
\title{Calculate the overall population immunity (aggregating over age)
from the supplied \code{popim_population} object}
\usage{
calc_pop_immunity(pop)
}
\arguments{
\item{pop}{\code{popim_population} object for which the population size and
immunity will be aggregated over age.}
}
\value{
dataframe containing the popim_population aggregated by age.
}
\description{
Calculate the overall population immunity (aggregating over age)
from the supplied \code{popim_population} object
}
\author{
Tini Garske
}
