% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popim_vacc_activities.R
\name{complete_vacc_activities}
\alias{complete_vacc_activities}
\title{Add coverage or doses information to the input
\code{popim_vacc_activities} object}
\usage{
complete_vacc_activities(vacc, pop)
}
\arguments{
\item{vacc}{\code{popim_vacc_activities} object}

\item{pop}{\code{popim_population} object}
}
\value{
The supplied object of class \code{popim_vacc_activities},
updated to have both \code{doses} and \code{coverage} information.
}
\description{
For each line in the \code{popim_vacc_activities} object the given
information of coverage is converted to doses, or vice versa,
using the target population size implied by the \code{popim_population}
object supplied. If both coverage and doses are given for any
activity, the function checks if they are consistent with the
population size, and fails if there are any inconsistencies.
}
\examples{
## set up some vaccination activities:
vacc <- popim_vacc_activities(region = "UK", year = 2001:2003,
                              age_first = 0, age_last = 0,
                              coverage = c(0.8, 0.8, NA),
                              doses = c(NA, NA, 60),
                              targeting = "random")
## set up a population to which these activities shall apply:
pop <- popim_population(region = "UK", year_min = 2000, year_max = 2005,
                        age_min = 0, age_max = 10)
pop$pop_size <- 100 ## cohort size of 100 for all cohorts

## fill in missing coverage/doses information based on population size:
vacc <- complete_vacc_activities(vacc, pop)

}
\author{
Tini Garske
}
