% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_vacc_activities.R
\name{complete_vacc_activities}
\alias{complete_vacc_activities}
\title{Adding coverage or doses information (whichever is missing) to the
vip_vacc_activities object}
\usage{
complete_vacc_activities(vaccs, pop_df)
}
\arguments{
\item{vaccs}{vip_vaccination_activities object}

\item{pop_df}{vip_population object}
}
\value{
vip_vaccination_activities object, updated to have both
doses and coverage information
}
\description{
For each line in the vip_vaccination_activities object the given
information of coverage is converted to doses, or vice versa,
using the target population size implied by the vip_population
object supplied. If both coverage and doses are given for any
activity, the function checks if they are consistent with the
population size, and fails if there are any inconsistencies.
}
\author{
Tini Garske
}
