% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_population.R
\name{is_population}
\alias{is_population}
\title{Function to test if the argument is a valid population dataframe}
\usage{
is_population(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{argument to be tested if it fulfils the requirements for
a population dataframe.}

\item{tol}{numerical, determines the tolerance when comparing two
numbers. Defaults to the square root of machine precision.}
}
\value{
logical
}
\description{
This function tests whether the supplied argument is a valid
population dataframe such as might be generated by the function
"vip_population".
}
\author{
Tini Garske
}
