% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{plot_immunity}
\alias{plot_pop_size}
\title{Plot the immunity or population size of a \code{popim_population} object}
\usage{
plot_immunity(pop)

plot_pop_size(pop)
}
\arguments{
\item{pop}{\code{popim_population} object such as created by
\code{\link[=popim_population]{popim_population()}}.}
}
\value{
A ggplot object.
}
\description{
Plot the immunity or population size of a \code{popim_population} object
}
\details{
The population is displayed in a grid showing the cohorts
through time. Time is shown on the x-axis, age on the y-axis,
such that a particular cohort tracks along diagonally from
bottom left to top right. If there are several regions, these
are shown as separate facets.

The colour in each cell corresponds to:
\itemize{
\item for \code{plot_immunity()}: the proportion of each cohort that is
immune, therefore varying between 0 and 1.
\item for \code{plot_pop_size()}: the size of each cohort.
}

As the returned object is a regular \code{ggplot} object, it can be
further modified with the ususal \code{ggplot2} syntax.
}
\author{
Tini Garske
}
