% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_population.R
\name{read_population}
\alias{read_population}
\title{Reading vip_population data from a .csv file}
\usage{
read_population(file)
}
\arguments{
\item{file}{Name of the .csv file from which the population data
are to be read. If it does not contain an absolute path, the
file name is relative to the current working directory.}
}
\value{
object of class "vip_population", a dataframe with one row
per birth cohort/year/region, with columns "region", "year",
"age", "cohort", "immunity", "pop_size".
}
\description{
Reads a population data from a .csv file, checks if
the data fulfils the requirements for a "vip_population"
object, and if so returns this object.
}
\details{
The requirements are that the data contain the columns \code{region},
\code{age}, \code{year} which will be coerced to character, integer,
integer. Columns \code{pop_size} and \code{immunity} are optional; they will
be coerced to numeric, and if missing will be initialised to 0 and
NA, respectively be set to NA and 0, respectively.

Limitations for values:
\itemize{
\item \code{age} must be non-negative integer
\item \code{immunity} must be between 0 and 1 (inclusive)
\item \code{pop_size} must be non-negative
}
}
\seealso{
\code{\link[utils:read.table]{utils::read.csv()}} which handles the reading of the .csv file.
}
\author{
Tini Garske
}
