% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip_population.R
\name{vip_population}
\alias{vip_population}
\title{Constructor of an object of the "vip_population" class}
\usage{
vip_population(
  region = character(),
  year_min = integer(),
  year_max = integer(),
  age_min = 0,
  age_max = 100
)
}
\arguments{
\item{region}{character vector, list of regions considered.}

\item{year_min}{integer, first year to be considered.}

\item{year_max}{integer, last year to be considered.}

\item{age_min}{integer, youngest age to be considered, defaults to
0. Must be non-negative.}

\item{age_max}{integer, oldest age to be considered, defaults to
100. Must be non-negative, and >= age_min.}
}
\value{
S3 object of class "vip_population": a dataframe with
columns region, year, age, cohort, immunity and pop_size,
where year and age cover the ranges given by the input
parameters. Cohort = year - age and gives the year of
birth. It is redundant but included for ease of
handling. Immunity and pop_size are initialised as 0 and NA,
respectively, throughout the whole population.
}
\description{
The "vip_population" object is a dataframe that models an
age-structured population through time, tracking population size
and vaccine-induced immunity in the population.
}
\details{
This constructor sets up the population as fully susceptible
(i.e., \code{immunity = 0}, with missing population size (i.e.,
\code{pop_size = NA_real_}) throughout. The parameters passed to the
constructor are retained as attributes to the dataframe object.

A population with non-missing population size can be read in from
a suitable file using \code{\link[=vip_pop_from_file]{vip_pop_from_file()}}, while vaccine induced
immunity can be generated through applying vaccination activities
to the population with \code{\link[=apply_vaccs]{apply_vaccs()}}.
}
\examples{
pop <- vip_population(region = "UK", year_min = 2000, year_max = 2010)

pop <- vip_population(region = c("FRA", "UK"),
                      year_min = 2000, year_max = 2010,
                      age_min = 0, age_max = 80)

}
\author{
Tini Garske
}
